Module hdk3::guest_callback

Modules

entry_defs

init
@todo figure out if/how we want to represent the init callback as a macro

migrate_agent
@todo figure out if/how we want to represent the migrate_agent callback as a macro

post_commit
@todo figure out if/how we want to represent the post_commit callback as a macro

validate
@todo figure out if/how we want to represent the validate callback as a macro

validate_link
@todo figure out if/how we want to represent the validate_link callback as a macro

validation_package
@todo figure out if/how we want to represent the validation_package callback as a macro