Struct hdk3::hash_path::path::Path

#[repr(transparent)]pub struct Path(_);

A Path is a vector of components. It represents a single traversal of a tree structure down to some arbitrary point. The main intent is that we can recursively walk back up the tree, hashing, committing and linking each sub-path along the way until we reach the root. At this point it is possible to follow DHT links from the root back up the path. i.e. the ahead-of-time predictability of the hashes of a given path allows us to travel "up" the tree and the linking functionality of the holochain DHT allows us to travel "down" the tree after at least one DHT participant has followed the path "up".

Implementations

impl Path

pub fn entry_def() -> EntryDef

pub fn entry_def_id() -> EntryDefId

pub fn entry_visibility() -> EntryVisibility

pub fn crdt_type() -> CrdtType

pub fn required_validations() -> RequiredValidations

impl Path

pub fn hash(&self) -> ExternResult<EntryHash>

What is the hash for the current Path?

pub fn exists(&self) -> ExternResult<bool>

Does an entry exist at the hash we expect?

pub fn ensure(&self) -> ExternResult<()>

Recursively touch this and every parent that doesn't exist yet.

pub fn parent(&self) -> Option<Path>

pub fn children(&self) -> ExternResult<Links>

Touch and list all the links from this anchor to anchors below it. Only returns links between anchors, not to other entries that might have their own links.

pub fn children_details(&self) -> ExternResult<LinkDetails>