Module hdk3::host_fn

Modules

agent_info

call

call_remote

create

create_link

debug

delete

delete_link

emit_signal
Emit an app-defined Signal.

entry_type_properties
@todo

get

get_agent_activity

get_details

get_link_details

get_links

property
@todo

query

random_bytes

remote_signal
Emit an app-defined Signal remotely on a list of agents.

schedule
@todo

show_env
@todo

sign

sys_time

unreachable
This is just a dummy module to flesh out the host_fn folder and make it mirror the ribosome. Also it does nothing and never will because it is the unreachable function that means a fn was called outside a callback that has permission to call it.

update

verify_signature

zome_info