Function hdk3::host_fn::call::call

pub fn call<I>(
    to_cell: Option<CellId>, 
    zome_name: ZomeName, 
    fn_name: FunctionName, 
    cap_secret: Option<CapSecret>, 
    payload: I
) -> ExternResult<ZomeCallResponse> where
    I: Serialize, 

Call

Make a Zome call in another Zome. The Zome can be in another Cell or the same Cell but must be installed on the same conductor.

Parameters

to_cell: The cell you want to call (If None will call the current cell).

zome_name: The name of the zome you want to call.

fn_name: The name of the function in the zome you are calling.

cap_secret: The capability secret if required.

payload: The arguments to the function you are calling.