Module hdk3::prelude::fixt

Fixturators for zome types

Structs

AgentPubKeyFixturator

AgentValidationPkgFixturator

AnyDhtHashFixturator

AppEntry
a curve to spit out Entry::App values

AppEntryBytesFixturator

AppEntryTypeFixturator

CapAccessFixturator

CapClaimFixturator

CapGrantFixturator

CapSecretFixturator

CloseChainFixturator

CrdtTypeFixturator

CreateFixturator

CreateLinkFixturator

CurryPayloadsFixturator

DeleteFixturator

DeleteLinkFixturator

DhtOpHashFixturator

DnaFixturator

DnaHashFixturator

EntryDefFixturator

EntryDefIdFixturator

EntryDefsFixturator

EntryFixturator

EntryHashFixturator

EntryTypeFixturator

EntryVisibilityFixturator

ExternIoFixturator

FunctionNameFixturator

GrantedFunctionFixturator

HashTypeAnyDhtFixturator

HeaderBuilderCommonFixturator

HeaderFixturator

HeaderHashFixturator

InitZomesCompleteFixturator

KnownCreateLink

KnownDeleteLink

LinkTagFixturator

MaybeSerializedBytesFixturator

MigrateAgentFixturator

NetIdHashFixturator

OpenChainFixturator

PublicCurve
A curve to make headers have public entry types

RequiredValidationTypeFixturator

RequiredValidationsFixturator

SignatureFixturator

ThirtySixHashBytesFixturator

TimestampFixturator

UpdateFixturator

WasmHashFixturator

ZomeCallCapGrantFixturator

ZomeIdFixturator

ZomeNameFixturator

Functions

element_with_no_entry

Type Definitions

HashTypeAnyDht

HashTypeEntry

MaybeSerializedBytes
Alias

ThirtySixHashBytes
A type alias for a Vec whose fixturator is expected to only return a Vec of length 36