Module hdk3::prelude::header

Structs

AgentValidationPkg
Header for an agent validation package, used to determine whether an agent is allowed to participate in this DNA

AppEntryType
Information about a class of Entries provided by the DNA

CloseChain
When migrating to a new version of a DNA, this header is committed to the old chain to declare the migration path taken. Currently unused

Create
A header which "speaks" Entry content into being. The same content can be referenced by multiple such headers.

CreateLink
Declares that a metadata Link should be made between two EntryHashes

Delete
Declare that a previously published Header should be nullified and considered deleted.

DeleteHeader
Placeholder for future when we want to have deletes on headers Not currently in use.

DeleteLink
Declares that a previously made Link should be nullified and considered removed.

Dna
The Dna Header is always the first header in a source chain

EntryDefIndex

HeaderBuilderCommon

HeaderHashedVec

HeaderHashes

InitZomesComplete
A header which declares that all zome init functions have successfully completed, and the chain is ready for commits. Contains no explicit data.

OpenChain
When migrating to a new version of a DNA, this header is committed to the new chain to declare the migration path taken. Currently unused

Update
A header which specifies that some new Entry content is intended to be an update to some old Entry.

UpdateHeader
Placeholder for future when we want to have updates on headers Not currently in use.

ZomeId
this id is an internal reference, which also serves as a canonical ordering for zome initialization. The value should be auto-generated from the Zome Bundle def

Enums

EntryType
Allows Headers which reference Entries to know what type of Entry it is referencing. Useful for examining Headers without needing to fetch the corresponding Entries.

Header
Header contains variants for each type of header.

HeaderType
A unit enum which just maps onto the different Header variants, without containing any extra data

Traits

HeaderBuilder
Builder for non-genesis Headers

HeaderInner
A trait to specify the common parts of a Header

Type Definitions

HeaderHashed