Enum hdk3::prelude::header::Header

pub enum Header {
    Dna(Dna),
    AgentValidationPkg(AgentValidationPkg),
    InitZomesComplete(InitZomesComplete),
    CreateLink(CreateLink),
    DeleteLink(DeleteLink),
    OpenChain(OpenChain),
    CloseChain(CloseChain),
    Create(Create),
    Update(Update),
    Delete(Delete),
}

Header contains variants for each type of header.

This struct really defines a local source chain, in the sense that it implements the pointers between hashes that a hash chain relies on, which are then used to check the integrity of data using cryptographic hash functions.

Variants

Dna(Dna)

AgentValidationPkg(AgentValidationPkg)

InitZomesComplete(InitZomesComplete)

CreateLink(CreateLink)

DeleteLink(DeleteLink)

OpenChain(OpenChain)

CloseChain(CloseChain)

Create(Create)

Update(Update)

Delete(Delete)

Implementations

impl Header

pub fn entry_data(&self) -> Option<(&HoloHash<Entry>, &EntryType)>

Returns the address and entry type of the Entry, if applicable.

pub fn entry_hash(&self) -> Option<&HoloHash<Entry>>

pub fn entry_type(&self) -> Option<&EntryType>

pub fn header_type(&self) -> HeaderType

pub fn author(&self) -> &HoloHash<Agent>

returns the public key of the agent who signed this header.

pub fn timestamp(&self) -> Timestamp

returns the timestamp of when the header was created

pub fn header_seq(&self) -> u32

returns the sequence ordinal of this header

pub fn prev_header(&self) -> Option<&HoloHash<Header>>

returns the previous header except for the DNA header which doesn't have a previous