Trait hdk3::prelude::holochain_serialized_bytes::serde::de::StdError

pub trait StdError: Debug + Display {
    fn source(&self) -> Option<&(dyn Error + 'static)> { ... }

    fn backtrace(&self) -> Option<&Backtrace> { ... }

    fn description(&self) -> &str { ... }

    fn cause(&self) -> Option<&dyn Error> { ... }
}

Error is a trait representing the basic expectations for error values, i.e., values of type E in Result<T, E>. Errors must describe themselves through the Display and Debug traits, and may provide cause chain information:

Error::source() is generally used when errors cross "abstraction boundaries". If one module must report an error that is caused by an error from a lower-level module, it can allow accessing that error via Error::source(). This makes it possible for the high-level module to provide its own errors while also revealing some of the implementation for debugging via source chains.

Provided methods

fn source(&self) -> Option<&(dyn Error + 'static)>

The lower-level source of this error, if any.

Examples

use std::error::Error;
use std::fmt;

#[derive(Debug)]
struct SuperError {
    side: SuperErrorSideKick,
}

impl fmt::Display for SuperError {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        write!(f, "SuperError is here!")
    }
}

impl Error for SuperError {
    fn source(&self) -> Option<&(dyn Error + 'static)> {
        Some(&self.side)
    }
}

#[derive(Debug)]
struct SuperErrorSideKick;

impl fmt::Display for SuperErrorSideKick {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        write!(f, "SuperErrorSideKick is here!")
    }
}

impl Error for SuperErrorSideKick {}

fn get_super_error() -> Result<(), SuperError> {
    Err(SuperError { side: SuperErrorSideKick })
}

fn main() {
    match get_super_error() {
        Err(e) => {
            println!("Error: {}", e);
            println!("Caused by: {}", e.source().unwrap());
        }
        _ => println!("No error"),
    }
}

fn backtrace(&self) -> Option<&Backtrace>

🔬 This is a nightly-only experimental API. (backtrace)

Returns a stack backtrace, if available, of where this error occurred.

This function allows inspecting the location, in code, of where an error happened. The returned Backtrace contains information about the stack trace of the OS thread of execution of where the error originated from.

Note that not all errors contain a Backtrace. Also note that a Backtrace may actually be empty. For more information consult the Backtrace type itself.

fn description(&self) -> &str

👎 Deprecated since 1.42.0:
use the Display impl or to_string()

if let Err(e) = "xc".parse::<u32>() {
    // Print `e` itself, no need for description().
    eprintln!("Error: {}", e);
}

fn cause(&self) -> Option<&dyn Error>

👎 Deprecated since 1.33.0:
replaced by Error::source, which can support downcasting

Loading content...
Implementations

impl dyn Error + 'static

pub fn is<T>(&self) -> bool where
T: 'static + Error,

Returns true if the boxed type is the same as T

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: 'static + Error,

Returns some reference to the boxed value if it is of type T, or None if it isn't.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: 'static + Error,

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

impl dyn Error + 'static + Send

pub fn is<T>(&self) -> bool where
T: 'static + Error,

Forwards to the method defined on the type dyn Error.

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: 'static + Error,

Forwards to the method defined on the type dyn Error.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: 'static + Error,

Forwards to the method defined on the type dyn Error.

impl dyn Error + 'static + Sync + Send

pub fn is<T>(&self) -> bool where
T: 'static + Error,

Forwards to the method defined on the type dyn Error.

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: 'static + Error,

Forwards to the method defined on the type dyn Error.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: 'static + Error,

Forwards to the method defined on the type dyn Error.

impl dyn Error + 'static

pub fn downcast<T>(
self: Box<dyn Error + 'static>
) -> Result<Box<T>, Box<dyn Error + 'static>> where
T: 'static + Error,

Attempts to downcast the box to a concrete type.

pub fn chain(&self) -> Chain<'_>

🔬 This is a nightly-only experimental API. (error_iter)

Returns an iterator starting with the current error and continuing with recursively calling Error::source.

If you want to omit the current error and only use its sources, use skip(1).

Examples

#![feature(error_iter)]
use std::error::Error;
use std::fmt;

#[derive(Debug)]
struct A;

#[derive(Debug)]
struct B(Option<Box<dyn Error + 'static>>);

impl fmt::Display for A {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        write!(f, "A")
    }
}

impl fmt::Display for B {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        write!(f, "B")
    }
}

impl Error for A {}

impl Error for B {
    fn source(&self) -> Option<&(dyn Error + 'static)> {
        self.0.as_ref().map(|e| e.as_ref())
    }
}

let b = B(Some(Box::new(A)));

// let err : Box<Error> = b.into(); // or
let err = &b as &(dyn Error);

let mut iter = err.chain();

assert_eq!("B".to_string(), iter.next().unwrap().to_string());
assert_eq!("A".to_string(), iter.next().unwrap().to_string());
assert!(iter.next().is_none());
assert!(iter.next().is_none());

impl dyn Error + 'static + Send

pub fn downcast<T>(
self: Box<dyn Error + 'static + Send>
) -> Result<Box<T>, Box<dyn Error + 'static + Send>> where
T: 'static + Error,

Attempts to downcast the box to a concrete type.

impl dyn Error + 'static + Sync + Send

pub fn downcast<T>(
self: Box<dyn Error + 'static + Sync + Send>
) -> Result<Box<T>, Box<dyn Error + 'static + Sync + Send>> where
T: 'static + Error,

Attempts to downcast the box to a concrete type.