Macro hdk3::prelude::debug_msg

macro_rules! debug_msg {
    ( $msg:expr ) => { ... };
    ( $msg:expr, $($tail:expr),* ) => { ... };
}

Returns a DebugMsg combining the message passed debug_msg! with the source code location in which it's called.

Examples

Basic usage

// Due to doc-test weirdness, this comment is technically on line 4.
let message: DebugMsg = debug_msg!("info: operation complete");

assert_eq!(message.msg(), "info: operation complete");
assert_eq!(message.file(), "src/debug.rs");
assert_eq!(message.line(), 5);

Advanced formatting

let operation = "frobnicate";

// Due to doc-test weirdness, this comment is technically on line 6.
let message: DebugMsg = debug_msg!(
    "info: operation complete: {}",
    operation
);

assert_eq!(message.msg(), "info: operation complete: frobnicate");
assert_eq!(message.file(), "src/debug.rs");
assert_eq!(message.line(), 7);