Struct hdk3::prelude::HoloHash

pub struct HoloHash<T> where
    T: HashType,  { /* fields omitted */ }

A HoloHash contains a vector of 36 bytes representing a 32-byte blake2b hash plus 4 bytes representing a DHT location. It also contains a zero-sized type which specifies what it is a hash of.

There is custom de/serialization implemented in [ser.rs]

Implementations

impl<T> HoloHash<T> where
T: HashType,

pub fn from_raw_39(hash: Vec<u8>) -> Result<HoloHash<T>, HoloHashError>

Raw constructor: Create a HoloHash from 39 bytes, using the prefix bytes to determine the hash_type

pub fn from_raw_39_panicky(hash: Vec<u8>) -> HoloHash<T>

Raw constructor: Create a HoloHash from 39 bytes, using the prefix bytes to determine the hash_type. Panics if hash_type does not match.

pub fn from_raw_36_and_type(bytes: Vec<u8>, hash_type: T) -> HoloHash<T>

Use a precomputed hash + location byte array in vec form, along with a type, to construct a hash. Used in this crate only, for testing.

pub fn retype<TT>(self, hash_type: TT) -> HoloHash<TT> where
TT: HashType,

Change the type of this HoloHash, keeping the same bytes

pub fn hash_type(&self) -> &T

The HashType of this hash

pub fn get_raw_39(&self) -> &[u8]

Get the raw 39-byte Vec including the 3 byte prefix, base 32 bytes, and the 4 byte loc

pub fn get_raw_36(&self) -> &[u8]

Get 36-byte Vec which excludes the 3 byte prefix

pub fn get_raw_32(&self) -> &[u8]

Fetch just the core 32 bytes (without the 4 location bytes)

pub fn get_loc(&self) -> u32

Fetch the holo dht location for this hash

pub fn into_inner(self) -> Vec<u8>

consume into the inner byte vector

impl<P> HoloHash<P> where
P: PrimitiveHashType,

pub fn from_raw_36(hash: Vec<u8>) -> HoloHash<P>

Construct from 36 raw bytes, using the known PrimitiveHashType

impl<T> HoloHash<T> where
T: HashType,

pub fn from_raw_32_and_type(hash: Vec<u8>, hash_type: T) -> HoloHash<T>

Construct a HoloHash from a 32-byte hash. The 3 prefix bytes will be added based on the provided HashType, and the 4 location bytes will be computed.

For convenience, 36 bytes can also be passed in, in which case the location bytes will used as provided, not computed.

impl<P> HoloHash<P> where
P: PrimitiveHashType,

pub fn from_raw_32(hash: Vec<u8>) -> HoloHash<P>

Construct a HoloHash from a prehashed raw 32-byte slice. The location bytes will be calculated.

impl<T> HoloHash<T> where
T: HashTypeSync,

pub fn with_data_sync<C>(content: &C) -> HoloHash<T> where
C: HashableContent<HashType = T>,

Synchronously hash a reference to the given content to produce a HoloHash If the content is larger than MAX_HASHABLE_CONTENT_LEN, this will panic!

impl<T> HoloHash<T> where
T: HashTypeAsync,

pub async fn with_data<C>(content: &'_ C) -> HoloHash<T> where
C: HashableContent<HashType = T>,

Asynchronously hash a reference to the given content to produce a HoloHash