Struct hdk3::prelude::SerializedBytes

#[repr(transparent)]pub struct SerializedBytes(_);

A Canonical Serialized Bytes representation for data If you have a data structure that needs a canonical byte representation use this Always round-trip through SerializedBytes via. a single TryFrom implementation. This ensures that the internal bytes of SerializedBytes are indeed canonical. The corrolary is that if bytes are NOT wrapped in SerializedBytes we can assume they are NOT canonical. Typically we need a canonical serialization when data is to be handled at the byte level by independently implemented and maintained systems.

Examples of times we need a canonical set of bytes to represent data:

cryptographic operations

moving across the host/guest wasm boundary

putting data on the network

Examples of times where we may not need a canonical representation and so may not need this:

round tripping data through a database that has its own serialization preferences

debug output or logging of data that is to be human readible

moving between data types within a single system that has no external facing representation

uses #[repr(transparent)] to maximise compatibility with ffi @see https://doc.rust-lang.org/1.26.2/unstable-book/language-features/repr-transparent.html#enter-reprtransparent

uses serde_bytes for efficient serialization and deserialization without this every byte will be individually round tripped through serde @see https://crates.io/crates/serde_bytes

Implementations

impl SerializedBytes

pub fn bytes(&self) -> &Vec<u8>