Trait hdk3::prelude::CallbackResult

pub trait CallbackResult {
    fn is_definitive(&self) -> bool;
}

Required methods

fn is_definitive(&self) -> bool

if a callback result is definitive we should halt any further iterations over remaining calls e.g. over sparse names or subsequent zomes typically a clear failure is definitive but success and missing dependencies are not in the case of success or missing deps, a subsequent callback could give us a definitive answer like a fail, and we don't want to over-optimise wasm calls and miss a clear failure

Loading content...