Module hdk3::prelude::x_salsa20_poly1305::key_ref

Structs

XSalsa20Poly1305KeyRef

Constants

KEY_REF_BYTES
Key refs are the same length as the keys themselves. The key ref is just a sha256 of the key. There are no benefits, only downsides, to having either a larger or smaller set of outputs (ref size) vs. the set of inputs (key size).

Type Definitions

SecretBoxKeyRef